Last commit

+++ src/main/java/com/example/postnews/web/response/PostFindAllResponse.java
@@ -22,7 +22,7 @@ public class PostFindAllResponse {
 
     private Instant updatedAt;
 
-    private Long countComment;
+    private Integer countComment;
 
 }
